Integrate feedback from sylvanc
Sylvan had feedback for the orignal code, so I integrated it.
- split into multiple files
- have Chain take an iterator of iterators
- change Limit to Take
- slight changes to the Cycle interface and implementation

Splitting into multiple files increases modularity.

Making Chain take an iterator of iterators means that it can now be
used to chain together an arbitrary number of iterators, instead of
having to chain together two at a time.

Take is a more common name than Limit for what that class does.

There were some issues with the generic types that were specified in
the Cycle interface. Also, it was storing things twice, once on
creation and once while it cycled through the iterator the first
time. The new implmentation takes advantage of `ArrayValues.rewind`,
which was added recently.